+++ src/Components/AxiosError/AxiosError.scss
@@ -22,6 +22,10 @@
     padding: 12% 40px 40px;
 }
 
+.error h3 {
+    text-align: center;
+}
+
 .error p {
     font-size: 20px;
     text-align: center;
