--- src/Components/NotFoudPage/NotFoudPage.scss
@@ -1,3 +1,0 @@
-.not-found {
-    height: 90vh;
-}
