+++ src/Components/NotFoundPage/NotFoundPage.scss
@@ -1,0 +1,25 @@
+.not-found {
+    height: 90vh;
+    padding: 10% 40px 40px;
+}
+
+.not-found__container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    align-content: center;
+    max-width: 600px;
+    margin: 0 auto;
+    border-radius: 10px
+}
+
+.not-found h2 {
+    font-size: 45px;
+    margin-bottom: 40px;
+    letter-spacing: 3px;
+}
+
+.not-found p {
+    font-size: 20px;
+    text-align: center;
+}
